Refactoring the codes.
