+++ crates/qasm3/src/exporter.rs
@@ -19,7 +19,10 @@ use crate::ast::{
     Reset, Statement, Version,
 };
 
+use crate::error::QASM3ExporterError;
 use crate::printer::BasicPrinter;
+use crate::symbol_table::SymbolTable;
+use crate::circuit_builder::CircuitBuilder;
 use hashbrown::{HashMap, HashSet};
 use indexmap::IndexMap;
 use pyo3::prelude::*;
@@ -32,7 +35,6 @@ use qiskit_circuit::circuit_data::CircuitData;
 use qiskit_circuit::operations::{DelayUnit, StandardInstruction};
 use qiskit_circuit::operations::{Operation, Param};
 use qiskit_circuit::packed_instruction::PackedInstruction;
-use thiserror::Error;
 
 use lazy_static::lazy_static;
 use regex::Regex;
@@ -40,22 +42,22 @@ use regex::Regex;
 type ExporterResult<T> = Result<T, QASM3ExporterError>;
 
 // These are the prefixes used for the loose qubit and bit names.
-const BIT_PREFIX: &str = "_bit";
-const QUBIT_PREFIX: &str = "_qubit";
+pub const BIT_PREFIX: &str = "_bit";
+pub const QUBIT_PREFIX: &str = "_qubit";
 
 // These are the prefixes used for the gate parameters and qubits.
-const GATE_PARAM_PREFIX: &str = "_gate_p";
-const GATE_QUBIT_PREFIX: &str = "_gate_q";
+pub const GATE_PARAM_PREFIX: &str = "_gate_p";
+pub const GATE_QUBIT_PREFIX: &str = "_gate_q";
 
 // These are the gates that are defined by the standard library.
-const GATES_DEFINED_BY_STDGATES: &[&str] = &[
+pub const GATES_DEFINED_BY_STDGATES: &[&str] = &[
     "p", "x", "y", "z", "h", "s", "sdg", "t", "tdg", "sx", "rx", "ry", "rz", "cx", "cy", "cz",
     "rzz", "cp", "crx", "cry", "crz", "ch", "swap", "ccx", "cswap", "cu", "CX", "phase",
     "cphase", "id", "u1", "u2", "u3",
 ];
 
 // These are the reserved keywords in QASM3.
-const RESERVED_KEYWORDS: &[&str] = &[
+pub const RESERVED_KEYWORDS: &[&str] = &[
     "OPENQASM",
     "angle",
     "array",
@@ -106,52 +108,39 @@ const RESERVED_KEYWORDS: &[&str] = &[
 ];
 
 lazy_static! {
-    static ref VALID_IDENTIFIER: Regex = Regex::new(r"(^[\w][\w\d]*$|^\$\d+$)").unwrap();
+    pub static ref VALID_IDENTIFIER: Regex = Regex::new(r"(^[\w][\w\d]*$|^\$\d+$)").unwrap();
 }
 
 lazy_static! {
-    static ref _BAD_IDENTIFIER_CHARACTERS: Regex = Regex::new(r"[^\w\d]").unwrap();
+    pub static ref _BAD_IDENTIFIER_CHARACTERS: Regex = Regex::new(r"[^\w\d]").unwrap();
 }
 
 lazy_static! {
-    static ref _VALID_HARDWARE_QUBIT: Regex = Regex::new(r"\$\d+").unwrap();
+    pub static ref _VALID_HARDWARE_QUBIT: Regex = Regex::new(r"\$\d+").unwrap();
 }
 
-#[derive(Error, Debug)]
-pub enum QASM3ExporterError {
-    #[error("Error: {0}")]
-    Error(String),
-    #[error("PyError: {0}")]
-    PyErr(PyErr),
-}
-
-impl From<PyErr> for QASM3ExporterError {
-    fn from(err: PyErr) -> Self {
-        QASM3ExporterError::PyErr(err)
-    }
-}
 
 #[derive(Debug, Eq, PartialEq, Hash, Clone)]
-enum BitType {
+pub enum BitType {
     Qubit(Qubit),
     Clbit(Clbit),
 }
 
 #[derive(Debug, Eq, PartialEq, Hash, Clone)]
-enum RegisterType {
+pub enum RegisterType {
     QuantumRegister(QuantumRegister),
     ClassicalRegister(ClassicalRegister),
 }
 
 impl RegisterType {
-    fn name(&self) -> &str {
+    pub fn name(&self) -> &str {
         match self {
             RegisterType::QuantumRegister(q) => q.name(),
             RegisterType::ClassicalRegister(c) => c.name(),
         }
     }
 
-    fn bits(&self, circuit_data: &CircuitData) -> Vec<BitType> {
+    pub fn bits(&self, circuit_data: &CircuitData) -> Vec<BitType> {
         match self {
             RegisterType::QuantumRegister(quantum_register) => quantum_register
                 .bits()
@@ -170,12 +159,12 @@ impl RegisterType {
 }
 
 #[derive(Debug, Clone)]
-struct Counter {
+pub struct Counter {
     current: usize,
 }
 
 impl Counter {
-    fn new() -> Self {
+    pub fn new() -> Self {
         Self { current: 0 }
     }
 }
@@ -203,7 +192,7 @@ impl BuildScope {
         
         // Map all qubits in the circuit
         for i in 0..circuit_data.num_qubits() {
-            let qubit = Qubit(i as u32);
+            let qubit = Qubit(i.try_into().unwrap_or_else(|_| panic!("Qubit index {} exceeds u32 range", i)));
             bit_map.insert(
                 BitType::Qubit(qubit),
                 BitType::Qubit(qubit),
@@ -212,7 +201,7 @@ impl BuildScope {
 
         // Map all clbits in the circuit
         for i in 0..circuit_data.num_clbits() {
-            let clbit = Clbit(i as u32);
+            let clbit = Clbit(i.try_into().unwrap_or_else(|_| panic!("Clbit index {} exceeds u32 range", i)));
             bit_map.insert(
                 BitType::Clbit(clbit),
                 BitType::Clbit(clbit),
@@ -233,280 +222,6 @@ impl BuildScope {
     }
 }
 
-#[derive(Debug)]
-struct LocalSymbols {
-    symbols: HashMap<String, Identifier>,
-    bitinfo: HashMap<BitType, Expression>,
-    reginfo: HashMap<RegisterType, Expression>,
-}
-
-impl LocalSymbols {
-    fn new() -> Self {
-        Self {
-            symbols: HashMap::new(),
-            bitinfo: HashMap::new(),
-            reginfo: HashMap::new(),
-        }
-    }
-}
-
-struct SymbolTable {
-    scopes: Vec<LocalSymbols>,
-    gates: IndexMap<String, QuantumGateDefinition>,
-    stdgates: HashSet<String>,
-    _counter: Counter,
-}
-
-impl SymbolTable {
-    fn new() -> Self {
-        Self {
-            scopes: vec![LocalSymbols::new()],
-            gates: IndexMap::new(),
-            stdgates: HashSet::new(),
-            _counter: Counter::new(),
-        }
-    }
-
-    fn bind(&mut self, name: &str) -> ExporterResult<()> {
-        if let Some(scope) = self.scopes.last() {
-            if !scope.symbols.contains_key(name) {
-                self.bind_no_check(name);
-            }
-        } else {
-            return Err(QASM3ExporterError::Error(format!(
-                "Symbol table is empty, cannot bind '{}'",
-                name
-            )));
-        }
-
-        Ok(())
-    }
-
-    fn bind_no_check(&mut self, name: &str) {
-        let id = Identifier {
-            string: name.to_string(),
-        };
-        if let Some(scope) = self.scopes.last_mut() {
-            scope.symbols.insert(name.to_string(), id);
-        }
-    }
-
-    fn contains_name(&self, name: &str) -> bool {
-        if let Some(scope) = self.scopes.last() {
-            scope.symbols.contains_key(name)
-        } else {
-            false
-        }
-    }
-
-    fn symbol_defined(&self, name: &str) -> bool {
-        RESERVED_KEYWORDS.contains(&name)
-            || self.gates.contains_key(name)
-            || self
-                .scopes
-                .iter()
-                .rev()
-                .any(|scope| scope.symbols.contains_key(name))
-    }
-
-    fn can_shadow_symbol(&self, name: &str) -> bool {
-        self.scopes.last()
-            .map(|scope| !scope.symbols.contains_key(name))
-            .unwrap_or(true)
-            && !self.gates.contains_key(name)
-            && !RESERVED_KEYWORDS.contains(&name)
-    }
-
-    fn escaped_declarable_name(
-        &mut self,
-        mut name: String,
-        allow_rename: bool,
-        unique: bool,
-    ) -> ExporterResult<String> {
-        let name_allowed = if unique {
-            |n: &str, this: &SymbolTable| !this.symbol_defined(n)
-        } else {
-            |n: &str, this: &SymbolTable| this.can_shadow_symbol(n)
-        };
-        if allow_rename {
-            if !VALID_IDENTIFIER.is_match(&name) {
-                name = format!("_{}", _BAD_IDENTIFIER_CHARACTERS.replace_all(&name, "_"));
-            }
-            while !name_allowed(&name, self) {
-                name = format!("{}{}", name, self._counter.next().expect("Counter should never fail"));
-            }
-            return Ok(name);
-        }
-        if !VALID_IDENTIFIER.is_match(&name) {
-            return Err(QASM3ExporterError::Error(format!(
-                "cannot use '{}' as a name; it is not a valid identifier",
-                name
-            )));
-        }
-
-        if RESERVED_KEYWORDS.contains(&name.as_str()) {
-            return Err(QASM3ExporterError::Error(format!(
-                "cannot use the keyword '{}' as a variable name",
-                name
-            )));
-        }
-
-        if !name_allowed(&name, self) {
-            if self.gates.contains_key(name.as_str()) {
-                return Err(QASM3ExporterError::Error(format!(
-                    "cannot shadow variable '{}', as it is already defined as a gate",
-                    name
-                )));
-            }
-
-            for scope in self.scopes.iter().rev() {
-                if let Some(other) = scope.symbols.get(&name) {
-                    return Err(QASM3ExporterError::Error(format!(
-                        "cannot shadow variable '{}', as it is already defined as '{:?}'",
-                        name, other
-                    )));
-                }
-            }
-
-            return Err(QASM3ExporterError::Error(format!(
-                "internal error: could not locate unshadowable '{}'",
-                name
-            )));
-        }
-
-        Ok(name.to_string())
-    }
-
-    fn add_standard_library_gates(&mut self) {
-        for gate in GATES_DEFINED_BY_STDGATES.iter() {
-            self.stdgates.insert(gate.to_string());
-        }
-    }
-
-    fn set_bitinfo(&mut self, id: Expression, bit: BitType) {
-        if self.scopes.is_empty() {
-            self.scopes.push(LocalSymbols::new());
-        }
-        if let Some(scope) = self.scopes.last_mut() {
-            scope.bitinfo.insert(bit, id);
-        }
-    }
-
-    fn get_bitinfo(&self, bit: &BitType) -> Option<&Expression> {
-        for scope in self.scopes.iter().rev() {
-            if let Some(id) = scope.bitinfo.get(bit) {
-                return Some(id);
-            }
-        }
-        None
-    }
-
-    fn set_reginfo(&mut self, id: Expression, reg: RegisterType) {
-        if self.scopes.is_empty() {
-            self.scopes.push(LocalSymbols::new());
-        }
-        if let Some(scope) = self.scopes.last_mut() {
-            scope.reginfo.insert(reg, id);
-        }
-    }
-
-    fn register_gate(
-        &mut self,
-        op_name: String,
-        params_def: Vec<Identifier>,
-        qubits: Vec<Identifier>,
-        body: QuantumBlock,
-    ) -> ExporterResult<()> {
-        // Changing the name is not allowed when defining new gates.
-        let name = self.escaped_declarable_name(op_name.clone(), false, false)?;
-        let _ = self.bind(&name);
-        self.gates.insert(
-            op_name,
-            QuantumGateDefinition {
-                quantum_gate_signature: QuantumGateSignature {
-                    name: Identifier { string: name },
-                    qarg_list: qubits,
-                    params: Some(
-                        params_def
-                            .into_iter()
-                            .map(|id| {
-                                Expression::Parameter(Parameter {
-                                    obj: id.string.clone(),
-                                })
-                            })
-                            .collect(),
-                    ),
-                },
-                quantum_block: body,
-            },
-        );
-        Ok(())
-    }
-
-    fn register_bits(
-        &mut self,
-        name: String,
-        bit: &BitType,
-        allow_rename: bool,
-        allow_hardware_qubit: bool,
-    ) -> ExporterResult<Identifier> {
-        if allow_hardware_qubit && _VALID_HARDWARE_QUBIT.is_match(&name) {
-            if self.symbol_defined(&name) {
-                return Err(QASM3ExporterError::Error(format!(
-                    "internal error: cannot redeclare hardware qubit {}",
-                    name
-                )));
-            }
-        } else {
-            self.escaped_declarable_name(name.clone(), allow_rename, false)?;
-        }
-        let identifier = Identifier {
-            string: name.clone(),
-        };
-        let _ = self.bind(&name);
-        self.set_bitinfo(
-            Expression::Parameter(Parameter {
-                obj: identifier.string.clone(),
-            }),
-            (*bit).to_owned(),
-        );
-        Ok(identifier)
-    }
-
-    fn register_registers(
-        &mut self,
-        name: String,
-        register: &RegisterType,
-    ) -> ExporterResult<Identifier> {
-        let name = self.escaped_declarable_name(name.clone(), true, false)?;
-        let identifier = Identifier {
-            string: name.clone(),
-        };
-        let _ = self.bind(&name);
-        self.set_reginfo(
-            Expression::Parameter(Parameter {
-                obj: identifier.string.clone(),
-            }),
-            (*register).to_owned(),
-        );
-        Ok(identifier)
-    }
-
-    fn push_scope(&mut self) {
-        self.scopes.push(LocalSymbols::new());
-    }
-
-    fn pop_scope(&mut self) {
-        self.scopes.pop();
-    }
-
-    fn new_context(&mut self) -> Self {
-        let mut new_table = SymbolTable::new();
-        new_table.gates.clone_from(&self.gates);
-        new_table.stdgates.clone_from(&self.stdgates);
-        new_table
-    }
-}
 
 pub struct Exporter {
     includes: Vec<String>,
@@ -635,14 +350,14 @@ impl<'a> QASM3Builder {
         let mut new_bit_map = HashMap::new();
 
         for i in 0..circuit_data.num_qubits() {
-            let qubit = Qubit(i as u32);
+            let qubit = Qubit(i.try_into().unwrap_or_else(|_| panic!("Qubit index {} exceeds u32 range", i)));
             new_bit_map.insert(
                 BitType::Qubit(qubit),
                 BitType::Qubit(qubit),
             );
         }
         for i in 0..circuit_data.num_clbits() {
-            let clbit = Clbit(i as u32);
+            let clbit = Clbit(i.try_into().unwrap_or_else(|_| panic!("Clbit index {} exceeds u32 range", i)));
             new_bit_map.insert(
                 BitType::Clbit(clbit),
                 BitType::Clbit(clbit),
@@ -670,14 +385,14 @@ impl<'a> QASM3Builder {
         let mut bit_map = HashMap::new();
 
         for i in 0..body.num_qubits() {
-            let qubit = Qubit(i as u32);
+            let qubit = Qubit(i.try_into().unwrap_or_else(|_| panic!("Qubit index {} exceeds u32 range", i)));
             bit_map.insert(
                 BitType::Qubit(qubit),
                 BitType::Qubit(qubit),
             );
         }
         for i in 0..body.num_clbits() {
-            let clbit = Clbit(i as u32);
+            let clbit = Clbit(i.try_into().unwrap_or_else(|_| panic!("Clbit index {} exceeds u32 range", i)));
             bit_map.insert(
                 BitType::Clbit(clbit),
                 BitType::Clbit(clbit),
@@ -714,8 +429,16 @@ impl<'a> QASM3Builder {
         let header = self.build_header();
 
         self.hoist_global_params()?;
-        let classical_decls = self.hoist_classical_bits()?;
-        let qubit_decls = self.build_qubit_decls()?;
+        
+        let mut circuit_builder = CircuitBuilder::new(
+            &self.circuit_scope.circuit_data,
+            &mut self.symbol_table,
+            self.is_layout,
+            self.allow_aliasing,
+        );
+        let classical_decls = circuit_builder.build_classical_declarations()?;
+        let qubit_decls = circuit_builder.build_qubit_declarations()?;
+        
         let main_stmts = self.build_top_level_stmts()?;
 
         let mut all_stmts = Vec::new();
@@ -790,253 +513,8 @@ impl<'a> QASM3Builder {
         })
     }
 
-    fn hoist_classical_bits(&mut self) -> ExporterResult<Vec<Statement>> {
-        let num_clbits = self.circuit_scope.circuit_data.num_clbits();
-        
-        let mut has_multiple_registers = false;
-        let registers: Vec<_> = self.circuit_scope.circuit_data.cregs().to_vec();
-        let mut bit_register_count = vec![0; num_clbits];
-        for register in &registers {
-            for shareable_clbit in register.bits() {
-                if let Some(clbit_index) = self.circuit_scope.circuit_data.clbits().find(&shareable_clbit) {
-                    bit_register_count[clbit_index.0 as usize] += 1;
-                    if bit_register_count[clbit_index.0 as usize] > 1 {
-                        has_multiple_registers = true;
-                        break;
-                    }
-                }
-            }
-            if has_multiple_registers {
-                break;
-            }
-        }
 
-        let mut decls = Vec::new();
-        if has_multiple_registers {
-            if !self.allow_aliasing {
-                return Err(QASM3ExporterError::Error(
-                    "Some Classical registers in this circuit overlap.".to_string(),
-                ));
-            }
-            for i in 0..num_clbits {
-                let clbit = Clbit(i as u32);
-                let identifier = self.symbol_table.register_bits(
-                    format!("{}{}", BIT_PREFIX, i),
-                    &BitType::Clbit(clbit),
-                    true,
-                    false,
-                )?;
-                decls.push(Statement::ClassicalDeclaration(ClassicalDeclaration {
-                    type_: ClassicalType::Bit,
-                    identifier,
-                }));
-            }
-            let registers: Vec<_> = self.circuit_scope.circuit_data.cregs().to_vec();
-            for register in registers {
-                let aliased =
-                    self.build_aliases(&RegisterType::ClassicalRegister(register.clone()))?;
-                decls.push(Statement::Alias(aliased));
-            }
-            return Ok(decls);
-        }
-        
-        let mut clbits_in_registers = HashSet::new();
-        for creg in self.circuit_scope.circuit_data.cregs() {
-            for shareable_clbit in creg.bits() {
-                if let Some(clbit_index) = self.circuit_scope.circuit_data.clbits().find(&shareable_clbit) {
-                    clbits_in_registers.insert(clbit_index.0);
-                }
-            }
-        }
 
-        // Declare loose clbits (not in any register)
-        for i in 0..num_clbits {
-            if !clbits_in_registers.contains(&(i as u32)) {
-                let clbit = Clbit(i as u32);
-                let identifier = self.symbol_table.register_bits(
-                    format!("{}{}", BIT_PREFIX, i),
-                    &BitType::Clbit(clbit),
-                    true,
-                    false,
-                )?;
-
-                decls.push(Statement::ClassicalDeclaration(ClassicalDeclaration {
-                    type_: ClassicalType::Bit,
-                    identifier,
-                }));
-            }
-        }
-        
-        // Declare registers
-        for creg in self.circuit_scope.circuit_data.cregs() {
-            let identifier = self.symbol_table.register_registers(
-                creg.name().to_string(),
-                &RegisterType::ClassicalRegister(creg.clone()),
-            )?;
-
-            for (i, shareable_clbit) in creg.bits().enumerate() {
-                if let Some(clbit) = self.circuit_scope.circuit_data.clbits().find(&shareable_clbit) {
-                    self.symbol_table.set_bitinfo(
-                        Expression::Index(Index {
-                            target: Box::new(Expression::Parameter(Parameter {
-                                obj: identifier.string.to_string(),
-                            })),
-                            index: Box::new(Expression::IntegerLiteral(IntegerLiteral(i as i64))),
-                        }),
-                        BitType::Clbit(clbit),
-                    )
-                }
-            }
-            decls.push(Statement::ClassicalDeclaration(ClassicalDeclaration {
-                type_: ClassicalType::BitArray(BitArray(creg.len() as u64)),
-                identifier,
-            }))
-        }
-        Ok(decls)
-    }
-
-    fn build_qubit_decls(&mut self) -> ExporterResult<Vec<Statement>> {
-        let num_qubits = self.circuit_scope.circuit_data.num_qubits();
-        
-        let mut has_multiple_registers = false;
-        let registers: Vec<_> = self.circuit_scope.circuit_data.qregs().to_vec();
-        let mut bit_register_count = vec![0; num_qubits];
-        for register in &registers {
-            for shareable_qubit in register.bits() {
-                if let Some(qubit_index) = self.circuit_scope.circuit_data.qubits().find(&shareable_qubit) {
-                    bit_register_count[qubit_index.0 as usize] += 1;
-                    if bit_register_count[qubit_index.0 as usize] > 1 {
-                        has_multiple_registers = true;
-                        break;
-                    }
-                }
-            }
-            if has_multiple_registers {
-                break;
-            }
-        }
-
-        let mut decls: Vec<Statement> = Vec::new();
-
-        if self.is_layout {
-            for i in 0..num_qubits {
-                let qubit = Qubit(i as u32);
-                self.symbol_table.register_bits(
-                    format!("${}", i),
-                    &BitType::Qubit(qubit),
-                    false,
-                    true,
-                )?;
-            }
-            return Ok(decls);
-        }
-
-        if has_multiple_registers {
-            if !self.allow_aliasing {
-                return Err(QASM3ExporterError::Error(
-                    "Some quantum registers in this circuit overlap.".to_string(),
-                ));
-            }
-
-            for i in 0..num_qubits {
-                let qubit = Qubit(i as u32);
-                let identifier = self.symbol_table.register_bits(
-                    format!("{}{}", if self.is_layout { "$" } else { QUBIT_PREFIX }, i),
-                    &BitType::Qubit(qubit),
-                    true,
-                    false,
-                )?;
-                decls.push(Statement::QuantumDeclaration(QuantumDeclaration {
-                    identifier,
-                    designator: None,
-                }));
-            }
-            let registers: Vec<_> = self.circuit_scope.circuit_data.qregs().to_vec();
-            for register in registers {
-                let aliased =
-                    self.build_aliases(&RegisterType::QuantumRegister(register.clone()))?;
-                decls.push(Statement::Alias(aliased));
-            }
-            return Ok(decls);
-        }
-
-        let mut qubits_in_registers = HashSet::new();
-        for qreg in self.circuit_scope.circuit_data.qregs() {
-            for shareable_qubit in qreg.bits() {
-                if let Some(qubit_index) = self.circuit_scope.circuit_data.qubits().find(&shareable_qubit) {
-                    qubits_in_registers.insert(qubit_index.0);
-                }
-            }
-        }
-
-        for i in 0..num_qubits {
-            if !qubits_in_registers.contains(&(i as u32)) {
-                let qubit = Qubit(i as u32);
-                let identifier = self.symbol_table.register_bits(
-                    format!("{}{}", if self.is_layout { "$" } else { QUBIT_PREFIX }, i),
-                    &BitType::Qubit(qubit),
-                    true,
-                    false,
-                )?;
-                decls.push(Statement::QuantumDeclaration(QuantumDeclaration {
-                    identifier,
-                    designator: None,
-                }));
-            }
-        }
-        
-        for qreg in self.circuit_scope.circuit_data.qregs() {
-            let identifier = self.symbol_table.register_registers(
-                qreg.name().to_string(),
-                &RegisterType::QuantumRegister(qreg.clone()),
-            )?;
-            for (i, shareable_qubit) in qreg.bits().enumerate() {
-                if let Some(qubit) = self.circuit_scope.circuit_data.qubits().find(&shareable_qubit) {
-                    self.symbol_table.set_bitinfo(
-                        Expression::Index(Index {
-                            target: Box::new(Expression::Parameter(Parameter {
-                                obj: identifier.string.to_string(),
-                            })),
-                            index: Box::new(Expression::IntegerLiteral(IntegerLiteral(i as i64))),
-                        }),
-                        BitType::Qubit(qubit),
-                    )
-                }
-            }
-            decls.push(Statement::QuantumDeclaration(QuantumDeclaration {
-                identifier,
-                designator: Some(Designator {
-                    expression: Expression::IntegerLiteral(IntegerLiteral(qreg.len() as i64)),
-                }),
-            }))
-        }
-        Ok(decls)
-    }
-
-    fn build_aliases(&mut self, register: &RegisterType) -> ExporterResult<Alias> {
-        let name = self
-            .symbol_table
-            .register_registers(register.name().to_string(), register)?;
-        let mut elements = Vec::new();
-        for (i, bit) in register.bits(&self.circuit_scope.circuit_data).iter().enumerate() {
-            let id = {
-                let temp_id = self.lookup_bit(bit)?;
-                temp_id.clone()
-            };
-            let id2 = Expression::Index(Index {
-                target: Box::new(Expression::Parameter(Parameter {
-                    obj: name.string.clone(),
-                })),
-                index: Box::new(Expression::IntegerLiteral(IntegerLiteral(i as i64))),
-            });
-            self.symbol_table.set_bitinfo(id2, bit.clone());
-            elements.push(id.clone());
-        }
-        Ok(Alias {
-            identifier: name,
-            value: Expression::IndexSet(IndexSet { values: elements }),
-        })
-    }
 
     fn build_top_level_stmts(&mut self) -> ExporterResult<Vec<Statement>> {
         let mut stmts = Vec::new();
